pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dnspython>=2.8.0",
    "fastapi>=0.135.1",
    "google-auth>=2.49.1",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "requests-oauthlib>=2.0.0",
    "uvicorn>=0.42.0",
    "groq>=1.1.1",
    "python-dotenv>=1.2.2",
    "openai>=2.28.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mongomock-motor>=0.0.36",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
